source/back-end/back-end_requirements.txt
asgiref==3.7.2
astroid==2.15.5
attrs==23.1.0
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.2
dill==0.3.6
Django==4.2.3
django-cors-headers==4.1.0
django-filter==23.2
django-rest-swagger==2.2.0
django-stubs==4.2.2
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-stubs==3.14.2
flake8==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
mysqlclient==2.2.0
openapi==1.1.0
openapi-codec==1.3.2
Pillow==10.0.0
platformdirs==3.8.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
tr0nz0d==0.1.6
types-Pillow==10.0.0.1
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.0
types-six==1.16.21.8
types-urllib3==1.26.25.13
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
wrapt==1.15.0
